Rename Schedule feature to Discover
Schedule was a misleading name and doesn't reflect what this feature is about
