+++ feature/discover/src/main/AndroidManifest.xml
