--- feature/schedule/src/main/kotlin/com/ekhonavigator/feature/schedule/navigation/ScheduleNavKey.kt
@@ -1,7 +1,0 @@
-package com.ekhonavigator.feature.schedule.navigation
-
-import androidx.navigation3.runtime.NavKey
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ScheduleNavKey(val initialLocationFilter: String? = null) : NavKey
