+++ settings.gradle.kts
@@ -38,7 +38,7 @@ include(":core:testing")
 include(":feature:account")
 include(":feature:calendar")
 include(":feature:home")
-include(":feature:schedule")
+include(":feature:discover")
 include(":feature:event")
 include(":feature:social")
 include(":feature:map")
